+++ src/main/java/finalmission/reservation/ui/ReservationController.java
@@ -1,10 +1,10 @@
 package finalmission.reservation.ui;
 
+import finalmission.auth.intercepter.AuthenticationPrincipal;
+import finalmission.auth.user.MemberInfo;
 import finalmission.reservation.application.ReservationCommandService;
-import finalmission.reservation.domain.MailClient;
 import finalmission.reservation.domain.Reservation;
 import finalmission.reservation.domain.ReservationRepository;
-import finalmission.reservation.domain.vo.ReservationApproval;
 import finalmission.reservation.ui.dto.ReservationRequest;
 import finalmission.reservation.ui.dto.ReservationResponse;
 import finalmission.reservation.ui.dto.ReservationUpdateRequest;
@@ -14,7 +14,6 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PatchMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
@@ -28,13 +27,11 @@ import org.springframework.web.bind.annotation.RestController;
 @RequestMapping(ReservationController.BASE_PATH)
 public class ReservationController {
 
-    public static final String BASE_PATH = "/reservation";
+    public static final String BASE_PATH = "/reservations";
 
     private final ReservationRepository reservationRepository;
 
     private final ReservationCommandService reservationCommandService;
-
-    private final MailClient mailClient;
 
     @GetMapping
     @ResponseStatus(HttpStatus.OK)
@@ -65,25 +62,10 @@ public class ReservationController {
         reservationCommandService.cancel(id);
     }
 
-    @GetMapping("/crews/{id}")
+    @GetMapping("/crews")
     @ResponseStatus(HttpStatus.OK)
     @Operation(summary = "크루의 모든 예약 조회")
-    public List<ReservationResponse> getAllByCrew(@PathVariable Long id) {
-        return reservationRepository.getAllByCrew(id);
-    }
-
-    @GetMapping("/coach/{id}")
-    @ResponseStatus(HttpStatus.OK)
-    @Operation(summary = "코치의 모든 예약 조회")
-    public List<ReservationResponse> getAllByCoach(@PathVariable Long id) {
-        return reservationRepository.getAllByCoach(id);
-    }
-
-    @PatchMapping("/{id}")
-    @ResponseStatus(HttpStatus.OK)
-    @Operation(summary = "예약 승인")
-    public void approval(@PathVariable Long id) {
-        ReservationApproval approval = reservationCommandService.approval(id);
-        mailClient.send(approval);
+    public List<ReservationResponse> getAllByCrew(@AuthenticationPrincipal MemberInfo memberInfo) {
+        return reservationRepository.getAllByCrew(memberInfo.id());
     }
 }
