+++ src/test/java/finalmission/FinalMissionApplicationTests.java
@@ -3,11 +3,11 @@ package finalmission;
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
-@SpringBootTest
+@SpringBootTest(classes = FinalMissionApplicationTests.class)
 class FinalMissionApplicationTests {
 
-  @Test
-  void contextLoads() {
-  }
+    @Test
+    void contextLoads() {
+    }
 
 }
